requirements.txt
langchain
langchain-openai
fastapi
python-dotenv
langfuse
requests
pytest